[bigint] Move the magnitude to 64-bit limbs
`BigInt` held its magnitude in base 2^32 while doing all its arithmetic
in 64-bit registers, so schoolbook multiplication and Knuth D both did
twice the passes they needed to. The words are the limbs now.

The kernels:

- `_add_word_pairs` / `_subtract_word_pairs` / `_submul_word_pairs`
  read two words as one limb; they are now `_add_words`,
  `_subtract_words` and `_submul_words`, one word per step, and every
  caller loses its odd-length tail.
- The packed base-2^64 schoolbook and the unpacked 32-bit one collapse
  into a single Comba kernel. `CUTOFF_PACK_64_AREA` is gone: there is
  nothing left to pack. Each column still splits its products at the
  word boundary into two `UInt128` accumulators, which is what keeps a
  column of 64-bit products from needing a three-word accumulator.
- Knuth D's quotient estimate takes the Moller-Granlund reciprocal
  after all. At a 32-bit limb it was pure loss, measured twice --
  64-by-32 is one `UDIV`. At 64 bits there is no 128-by-64 instruction
  on either target, so the alternative is a software helper and the
  reciprocal is not optional. `_divmod_single_word()` uses it too,
  scaling the dividend a word at a time since it cannot scale it once.

Decimal conversion goes to `10^18` chunks. Nineteen digits would fit a
word, but eighteen is exactly twice what `10^9` carried in a half-as-
wide word -- so the density per byte is unchanged and the nine-digit
grouping survives, which `to_biguint()` needs, since it hands the
chunks to `BigUInt` and only whole nine-digit groups split cleanly.

Four things were wrong in ways a type error could not catch, and each
is a place where 32-bit arithmetic was correct for a reason that stops
holding at 64:

1. `to_biguint()` reinterpreted the chunks as `BigUInt` words, which is
   why the chunk base cannot be `10^19`.
2. `String(Int(words[0]))` formatted a word through a *signed* type.
   Every value at or above 2^63 printed negative: 2^64 - 1 came out
   as -1.
3. `borrow = (diff >> 63) & 1` read the borrow out of the sign bit of
   a widened 32-bit subtraction. At 64 bits a word of exactly 2^63
   sets that bit with no borrow at all. It is `word < borrow` now.
4. Toom-3's exact division by three carried `(2^32 - 1) / 3`, and its
   halving shifted the carry in at bit 31.

`_sqrtrem_two_words()` needed a real 128-bit root, since two words is
no longer a `UInt64`; `_isqrt_uint128()` descends by Newton from
`(isqrt(high) + 1) * 2^32`, which is at or above the answer because
`value < (high + 1) * 2^64`.

The NTT packs from 64-bit words now, but `_unpack()` still walks 32
bits at a time: its accumulator holds a coefficient shifted by up to
the step width, and stepping a whole word would both double the number
of terms and shift them twice as far, which no longer fits 128 bits.

`WordList` keeps the same 48 bytes inline -- twelve 32-bit words became
six 64-bit ones, the same 384 bits.

Checked: multiplication verified by `(a+b)^2 - (a-b)^2 == 4ab` from 10
digits to 10^6, across schoolbook, Karatsuba, Toom-3 and the transform;
division by `q*b + r == a` over 759 operand pairs; sqrt by
`s*s <= x < (s+1)^2` over 284; decimal round-trip across every word
boundary. Cutoffs are not re-swept yet -- a word is twice as wide, so
every one of them means a different size now.

+++ src/decimo/bigint/special.mojo
@@ -120,9 +120,9 @@ def product_range(low: Int, high: Int) raises -> BigInt:
     # reserve the result up front to avoid reallocating while it grows.
     if high - low + 1 <= PRODUCT_RANGE_LEAF_CUTOFF:
         var result = BigInt(uninitialized_capacity=high - low + 2)
-        result.words.append(UInt32(low))
+        result.words.append(UInt64(low))
         for factor in range(low + 1, high + 1):
-            multiply_by_word_inplace(result, UInt32(factor))
+            multiply_by_word_inplace(result, UInt64(factor))
         return result^
     var mid = low + (high - low) // 2
     return product_range(low, mid) * product_range(mid + 1, high)
